+++ src/index.js
@@ -26,16 +26,8 @@ function mergeStrategy(rules = {}) {
   return function () {
     return lodashMerge.apply(null, [{}].concat(...arguments).concat([
       joinArrays({
-        customizeArray: (a, b, key) => {
-          if (rules[key]) {
-            return rules[key] === 'prepend' && b.concat(a);
-          }
-        },
-        customizeObject: (a, b, key) => {
-          if (rules[key]) {
-            return rules[key] === 'prepend' && lodashMerge({}, b, a, joinArrays());
-          }
-        }
+        customizeArray: customizeArray(rules),
+        customizeObject: customizeObject(rules)
       })
     ]));
   };
@@ -44,29 +36,65 @@ function mergeStrategy(rules = {}) {
 function mergeSmartStrategy(rules = {}) {
   // rules: { <field>: <'append'|'prepend'> }
   // All default to append but you can override here
+  const customArray = customizeArray(rules);
+
   return function () {
     return lodashMerge.apply(null, [{}].concat(...arguments).concat([
       joinArrays({
         customizeArray: (a, b, key) => {
           if (isLoader(key.split('.').slice(-1)[0])) {
-            if (rules[key]) {
-              return b.reduce(reduceLoaders, a.slice());
+            const rule = rules[key];
+
+            if (rule) {
+              if (rule === 'prepend') {
+                return b.reduce(reduceLoaders, a.slice());
+              }
+
+              if (rule === 'replace') {
+                return b;
+              }
             }
 
             return a.reduce(reduceLoaders, b.slice());
           }
 
-          if (rules[key]) {
-            return rules[key] === 'prepend' && b.concat(a);
-          }
+          return customArray(a, b, key);
         },
-        customizeObject: (a, b, key) => {
-          if (rules[key]) {
-            return rules[key] === 'prepend' && lodashMerge({}, b, a, joinArrays());
-          }
-        }
+        customizeObject: customizeObject(rules)
       })
     ]));
+  };
+}
+
+function customizeArray(rules) {
+  return (a, b, key) => {
+    const rule = rules[key];
+
+    if (rule) {
+      if (rule === 'prepend') {
+        return b.concat(a);
+      }
+
+      if (rule === 'replace') {
+        return b;
+      }
+    }
+  };
+}
+
+function customizeObject(rules) {
+  return (a, b, key) => {
+    const rule = rules[key];
+
+    if (rule) {
+      if (rule === 'prepend') {
+        return lodashMerge({}, b, a, joinArrays());
+      }
+
+      if (rule === 'replace') {
+        return b;
+      }
+    }
   };
 }