+++ tests/test-merge-smart-strategy.js
@@ -54,4 +54,44 @@ function mergeStrategySpecificTests(merge) {
       'module.loaders': 'prepend'
     })(a, b), result);
   });
+
+  it('should work with nested arrays and replace', function () {
+    const a = {
+      module: {
+        loaders: [
+          {
+            test: /.jsx?$/,
+            loaders: ['babel'],
+            exclude: /node_modules/
+          }
+        ]
+      }
+    };
+    const b = {
+      module: {
+        loaders: [
+          {
+            test: /.jsx?$/,
+            loaders: ['react-hot'],
+            exclude: /node_modules/
+          }
+        ]
+      }
+    };
+    const result = {
+      module: {
+        loaders: [
+          {
+            test: /.jsx?$/,
+            loaders: ['react-hot'],
+            exclude: /node_modules/
+          }
+        ]
+      }
+    };
+
+    assert.deepEqual(merge({
+      'module.loaders': 'replace'
+    })(a, b), result);
+  });
 }